AP-61 | Move the action into separate components and add a snapshot test for each of them
