+++ ui/react-components/components/ListViewAction/ListViewAction.jsx
@@ -1,0 +1,9 @@
+import {useState} from "react";
+
+export default function useListViewAction() {
+  const [show, setShow] = useState(true);
+
+  const handleClose = () => setShow(false);
+
+  return [show, handleClose]
+}
